Restaurant system
 Restaurant system: - The restaurant is divided into two parts: one side is managed by the staff taking
orders and the other is managed by the manager who manages the staff, the menu
content, and can make salary reports. - The cashier employee takes customer orders and information to complete the order
through the system and registers customers who want to earn points to make a
discount after a while of orders. - The customer information will assist us in sending promos and points to them. - The invoice will have information of the order like the date of order, how many
items have been ordered and information about customers with points they earned,
and so on  - The manager can access sales reports which can also know the most order items, the
least order items, and the total orders. The manager can also modify the menu, and it
is possible to add items and set a price for each item, or it is possible to delete items.